Check if data set is empty in plotter.py
Python: plotter.py: Print error message if the data set is empty, but exit with status code 0, because this is not technically an error. But still print to stderr, because we pipe stdout to /dev/null in the Ruby code.
Also use sys.exit properly in plotter.py.
It defaults to exiting with 0 if no argument is supplied, which is bad.