Fix bugs sorting ConfigWithTotalConfusionMatrix
Fix bugs sorting ConfigWithTotalConfusionMatrix.
Correct the sorting, by using operator> rather than operator<, so that
the more points are awarded to a configuration the earlier it appears in
the list.
Prevent underflows from occurring when the negative score is greater
than the positive, in that case award 0 points (the worst possible
rating) to the configuration.